ueventd: make parallel restorecon functionality optional
cdcc3469b5a867dbf04e3203d7e7f45a5fb9cfb5 [formerly 5aa6197d5f387579ff04c330001840d6988e825f] added the ability to
parallelize restorecon to speed up boot for devices that have not
completely moved to genfscon.  This parallel restorecon happens after
the parallel ueventd handling.

This causes a performance regression for devices that have moved to
genfscon, since previously, the restorecon() was done in the main
ueventd thread in parallel with the uevent handlers.

I also tried to run the fully parallelized restorecon in parallel with
the uevent handlers, but that did not make any change to the cold boot
time, likely due to the additional overhead of parallelizing the work.

Bug: 140458170
Test: blueline coldboot time returns to pre-regression time.
Change-Id: I3cd6a869cc9b62792466813d94ad6c69834e854e

Former-commit-id: 78c36854fe470e4f6f18286796a5b35719192013
Signed-off-by: Joey Huab <joey@evolution-x.org>

Former-commit-id: c5be6e0d4e980b1c3ff9c2da3eaec244208655a4