Change search order for system processes in vndk_lite.
Originally, vndk_lite does not include product/lib and product_services/lib
directories first but searches for the required files in vendor/lib and odm/lib.
However, if a vendor/odm library with the same name exists, that causes the
vendor/odm library to be set as dependency to the system process that is searching
for the library, hitting neverallow SELinux denials.
To avoid this problem, allow the system processes in vndk_lite devices
to search product/lib and product_services/lib prior to vendor/lib and odm/lib.

Test: Check boot for vndk_lite devices.

Change-Id: I4113938d0803480f682c3521765e1d5f3c19ab79

Former-commit-id: 6f48b2ffe1663e0c590f5f5ff559e76f95408729
Signed-off-by: Joey Huab <joey@evolution-x.org>

Former-commit-id: 5eda253f1730aa6914e6a2cdcc9101dd29f18c9f